+++ src/zodbupdate/tests.py
@@ -14,10 +14,12 @@
 ##############################################################################
 
 import ZODB
+import ZODB.blob
 import ZODB.broken
 import ZODB.FileStorage
 import os
 import persistent
+import shutil
 import sys
 import tempfile
 import transaction
@@ -89,8 +91,12 @@ class Tests(unittest.TestCase):
         IOtherFactory.__module__ = 'module2.interfaces'
 
         self.tmphnd, self.dbfile = tempfile.mkstemp()
+        self.tmpblob = tempfile.mkdtemp()
 
-        self.storage = ZODB.FileStorage.FileStorage(self.dbfile)
+        self.storage = ZODB.blob.BlobStorage(
+            self.tmpblob,
+            ZODB.FileStorage.FileStorage(self.dbfile),
+        )
         self.db = ZODB.DB(self.storage)
         self.conn = self.db.open()
         self.root = self.conn.root()
@@ -102,12 +108,18 @@ class Tests(unittest.TestCase):
         self.db.close()
         self.storage.close()
 
-        self.storage = ZODB.FileStorage.FileStorage(self.dbfile)
+        self.storage = ZODB.blob.BlobStorage(
+            self.tmpblob,
+            ZODB.FileStorage.FileStorage(self.dbfile),
+        )
         updater = zodbupdate.main.create_updater(self.storage, **args)
         updater()
         self.storage.close()
 
-        self.storage = ZODB.FileStorage.FileStorage(self.dbfile)
+        self.storage = ZODB.blob.BlobStorage(
+            self.tmpblob,
+            ZODB.FileStorage.FileStorage(self.dbfile),
+        )
         self.db = ZODB.DB(self.storage)
         self.conn = self.db.open()
         self.root = self.conn.root()
@@ -134,6 +146,7 @@ class Tests(unittest.TestCase):
         os.unlink(self.dbfile + '.index')
         os.unlink(self.dbfile + '.tmp')
         os.unlink(self.dbfile + '.lock')
+        shutil.rmtree(self.tmpblob)
 
         zodbupdate.serialize.SKIP_SYMBS = self._skipped_symbs
 
@@ -184,17 +197,13 @@ class Tests(unittest.TestCase):
         skipped = sys.modules['module1'].Factory()
         self.root['skipped'] = skipped
         transaction.commit()
-
         self.assertIn(('ZODB.blob', 'Blob'), zodbupdate.serialize.SKIP_SYMBS)
         zodbupdate.serialize.SKIP_SYMBS += [('module1', 'Factory')]
-
         oid = self.root['skipped']._p_oid
         old_pickle, old_serial = self.storage.load(oid)
-
         self.update(
             default_renames={
                 ('module1', 'Factory'): ('module2', 'OtherFactory')})
-
         pickle, serial = self.storage.load(oid)
         self.assertEqual(old_pickle, pickle)
         self.assertEqual(old_serial, serial)
@@ -203,18 +212,14 @@ class Tests(unittest.TestCase):
         skipped = sys.modules['module1'].Factory()
         self.root['skipped'] = skipped
         transaction.commit()
-
         self.assertNotIn(
             ('module1', 'Factory'),
             zodbupdate.serialize.SKIP_SYMBS)
-
         oid = self.root['skipped']._p_oid
         old_pickle, old_serial = self.storage.load(oid)
-
         self.update(
             default_renames={
                 ('module1', 'Factory'): ('module2', 'OtherFactory')})
-
         pickle, serial = self.storage.load(oid)
         self.assertNotEqual(old_pickle, pickle)
         self.assertNotEqual(old_serial, serial)
@@ -678,6 +683,26 @@ class Python2Tests(Tests):
         result = encoder(mock)
         self.assertEquals(result, False)
         self.assertEquals(mock['foo'], None)
+
+    def test_blobs_are_left_untouched(self):
+        blob = ZODB.blob.Blob()
+        self.root['blob'] = blob
+        transaction.commit()
+
+        oid = self.root['blob']._p_oid
+        old_pickle, old_serial = self.storage.load(oid)
+
+        self.update(convert_py3=True)
+
+        pickle, serial = self.storage.load(oid)
+        self.assertEqual(old_pickle, pickle)
+        self.assertEqual(old_serial, serial)
+
+        # make sure the reference to the blob was converted
+        self.assertIn(
+            b'C\x08\x00\x00\x00\x00\x00\x00\x00\x01cZODB.blob',
+            self.storage.load(self.root._p_oid, '')[0]
+        )
 
 
 class Python3Tests(Tests):