+++ CHANGES.rst
@@ -8,7 +8,7 @@ Changes
    the new command line option ``--encoding-fallback``.
   (`#15 <https://github.com/zopefoundation/zodbupdate/pull/15>`_)
 
-- Switch to use `argparse` b/c `optparse` is deprecated.
+- Switch to use `argparse` as `optparse` is deprecated.
 
 - Add ability to run the Python 3 migration with a default encoding for
   ``str`` objects.